// src/sdk/types/enums.ts
export const enum HttpMethodsEnum {
    Get = "GET",
    Post = "POST",
    Put = "PUT",
    Delete = "DELETE",
}

export const enum MessageTypesEnum {
    Audio = "audio", // for audio messages.
    Contacts = "contacts", // for contact messages.
    Document = "document", // for document messages.
    Image = "image", // for image messages.
    Interactive = "interactive", // for list and reply button messages.
    Location = "location", // for location messages.
    Reaction = "sticker", // for reaction messages.
    Sticker = "sticker", // for sticker messages.
    Template = "template", // for template messages. Text and media (images and documents) message templates are supported.
    Text = "text", // for text messages.
    Video = "video", // for video messages
}
// ...
export const enum ComponentTypesEnum {
    Header = "header",
    Body = "body",
    Button = "button",
}

export enum WAConfigEnum {
    BaseURL = "WA_BASE_URL",
    AppId = "M4D_APP_ID",
    AppSecret = "M4D_APP_SECRET",
    PhoneNumberId = "WA_PHONE_NUMBER_ID",
    BusinessAcctId = "WA_BUSINESS_ACCOUNT_ID",
    APIVersion = "CLOUD_API_VERSION",
    AccessToken = "CLOUD_API_ACCESS_TOKEN",
    WebhookEndpoint = "WEBHOOK_ENDPOINT",
    WebhookVerificationToken = "WEBHOOK_VERIFICATION_TOKEN",
    ListenerPort = "LISTENER_PORT",
    MaxRetriesAfterWait = "MAX_RETRIES_AFTER_WAIT",
    RequestTimeout = "REQUEST_TIMEOUT",
    Debug = "DEBUG",
}